refactor(Kafka Trigger Node): Move the v2 consumer factory next to the producer (no-changelog)
Mirror the layout the v2 producer settled on: the factory that builds a client
and a consumer lives in transport, alongside producer.ts, and the barrel
exports it with its options type. The consume loop and the message parser stay
in v2/consumer, since unlike the producer node they cannot live in the node
file: the trigger's runtime has to be testable without one.

Rename the loop module to consume-topic.ts so it does not read as a second
consumer.ts, and split its tests to match.
