+++ packages/nodes-base/nodes/Kafka/v2/consumer/index.ts
@@ -1,15 +1,9 @@
-export {
-	CONSUMER_DEFAULTS,
-	DEFAULT_PARTITIONS_CONSUMED_CONCURRENTLY,
-	consumeTopic,
-	createKafkaConsumer,
-} from './consumer';
+export { consumeTopic, DEFAULT_PARTITIONS_CONSUMED_CONCURRENTLY } from './consume-topic';
 export type {
 	ConsumeTopicOptions,
 	KafkaBatchHandOff,
 	KafkaBatchHandler,
 	KafkaConsumerHandle,
-	KafkaConsumerOptions,
-} from './consumer';
+} from './consume-topic';
 export { createMessageParser } from './message-parser';
 export type { KafkaMessageParser, KafkaMessageParserOptions } from './message-parser';
