+++ packages/nodes-base/nodes/Kafka/v2/transport/index.ts
@@ -1,1 +1,4 @@
+export { toKafkaJSConfig } from './config';
+export { getKafkaLibrary } from './client';
 export { createKafkaProducer, type KafkaProducerOptions } from './producer';
+export { createKafkaConsumer, CONSUMER_DEFAULTS, type KafkaConsumerOptions } from './consumer';
